feat: extract mesh utilities into module
